configure jitpack build
The build was timing out after 20 minutes so we needed to make it a bit
leaner. Turns out our patch is on the compiler and we consume it via the
gradle plugin. The gradle plugin both handles compilation and adds the
runtime component of sqldelight as a dependency. When it adds the
runtime component, it does so via a hardcoded package dependecy. We are
ok with that since the runtime won't affect reproducibility.

So we can build only the compiler and gradle plugin, and be happy!

We also need to set the group so that the dependencies resolve
correctly.